Merge upstream faktory (#11)
* Add KeepAlive to TCP Socket

* Fix jobBatchId parser to work for all job types (#92)

Faktory seems to use both _bid and bid as keys for the batch id in the
custom object for jobs. Normal jobs get the 'bid' key and callback jobs
seem to get the '_bid' formulation.

This makes the parser able to parse both styles.

* Update README.lhs

* Update README.lhs

* General housekeeping

* Update resolvers
* Rewrite cabal file with newer stack
* Convert project to Fourmolu
* Use dev restylers
* Silence alias lint error

* Migrate to crypton-connection

* Use lists for ghc-options in package.yaml

hpack accepts either, and this allows it to [auto-]format the nicest.

* Drop support for lts-14 / ghc-8.6

Can't build crypton, can't use crypton-connection. Losing 8.6 isn't the
end of the world -- we still support 6 major versions (8.8 - 9.6).

* Version bump

* Bump nightly resolver

* Disable new warning in 9.8

* Minor CPP for aeson-2.2.0

* Re-organize resolvers

* Version bump

* Re-use withWorker

* Add WorkerSettings{settingsOnFailed}

This may be generally useful, but the immediate need is it will allow us
to test some behavior around it, which we are about to improve.

* Add failing test for parse errors

Reproduces #88.

* Fail Jobs that don't decode as expected

Originally, this decoding failure would enter the `Left` case and only
log the error. The Job would hang, neither `ACK`ed nor `FAIL`ed until
its reservation expired.

Now, as long as it is valid JSON, we reach the `Right`-`Just` branch and
attempt to decode it further there. This way, if it can't decode, that
exception occurs in the right place to `FAIL` the job, as desired.

Closes #88.

* Version bump

* Add Faktory.Pool

This implements connection pooling by adding `HasFaktoryPool` which can
be used to acquire a `Producer`, which is just a `Client`, which is just
a `Connection`.

I chose the name `FaktoryPool` (instead of `FaktoryProducerPool`) for
two reasons, one conceptual and one practical:

- Conceptual: I think the `Producer` wrapper over `Client` is
  unnecessary. It adds nothing and a `Producer` could just as easily be
  used for any communication with Faktory, to produce, consume, batch,
  track, etc. So I consider this a "pool for Faktory" and no
  specifically "for Faktory _producers_". (Same reason we call it
  `SqlPool` and not `SqlInserterPool` or `SqlQuerierPool`.)
- Practical: we have a `FaktoryProducerPool` in `freckle-app`, that this
  is a generalization and upstreaming of. Changing the name along the
  way makes the errors clearer and more easy to deal with when
  converting downstream applications.

* Version bump

* Use newtype instead of type

* Compatibility with resource-pool < 0.3

* update website url (#103)

* Fix import.

---------

Co-authored-by: patrick brisbin <pbrisbin@gmail.com>
Co-authored-by: AJ Ruf <z0isch@gmail.com>
Co-authored-by: Chris Martin <ch.martin@gmail.com>